requirements.txt
Babel==2.7.0
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-appconf==1.0.3
django-bootstrap-modal-forms==1.4.1
django-crispy-forms==1.7.2
django-environ==0.4.5
django-extensions==2.1.9
django-filter==2.2.0
django-import-export==1.2.0
django-mathfilters==0.4.0
django-select2==6.3.1
django-smart-selects==1.5.4
django-widget-tweaks==1.4.2
et-xmlfile==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.0.0
psycopg2==2.7.7
psycopg2-binary==2.8.2
pytz==2019.1
PyYAML==5.1.2
selenium==3.14.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
